setting for the installation directory

+++ wpmcpp/packageversion.cpp
@@ -261,8 +261,7 @@ void PackageVersion::uninstall(Job* job)
 
 QDir PackageVersion::getDirectory()
 {
-    QString pf = WPMUtils::getProgramFilesDir();
-    QDir d(pf + "\\WPM\\" + this->package + "-" +
+    QDir d(WPMUtils::getInstallationDirectory() + "\\" + this->package + "-" +
            this->version.getVersionString());
     return d;
 }
@@ -730,32 +729,34 @@ void PackageVersion::deleteShortcuts(QDir& d)
                 QDir dd(path);
                 deleteShortcuts(dd);
             } else {
-                IPersistFile* ppf;
+                if (path.toLower().endsWith(".lnk")) {
+                    IPersistFile* ppf;
 
-                // Query IShellLink for the IPersistFile interface for saving the
-                // shortcut in persistent storage.
-                hres = psl->QueryInterface(IID_IPersistFile, (LPVOID*)&ppf);
+                    // Query IShellLink for the IPersistFile interface for saving the
+                    // shortcut in persistent storage.
+                    hres = psl->QueryInterface(IID_IPersistFile, (LPVOID*)&ppf);
 
-                if (SUCCEEDED(hres)) {
-                    // Save the link by calling IPersistFile::Save.
-                    hres = ppf->Load((WCHAR*) path.utf16(), STGM_READ);
                     if (SUCCEEDED(hres)) {
-                        WCHAR info[MAX_PATH + 1];
-                        hres = psl->GetPath(info, MAX_PATH,
-                                (WIN32_FIND_DATAW*) 0, SLGP_UNCPRIORITY);
+                        // Save the link by calling IPersistFile::Save.
+                        hres = ppf->Load((WCHAR*) path.utf16(), STGM_READ);
                         if (SUCCEEDED(hres)) {
-                            QString targetPath;
-                            targetPath.setUtf16((ushort*) info, wcslen(info));
-                            // qDebug() << "deleteShortcuts " << targetPath << " " <<
-                            //        instPath;
-                            if (WPMUtils::isUnder(targetPath,
-                                                  instPath)) {
-                                QFile::remove(path);
-                                // qDebug() << "deleteShortcuts true" << ok;
+                            WCHAR info[MAX_PATH + 1];
+                            hres = psl->GetPath(info, MAX_PATH,
+                                    (WIN32_FIND_DATAW*) 0, SLGP_UNCPRIORITY);
+                            if (SUCCEEDED(hres)) {
+                                QString targetPath;
+                                targetPath.setUtf16((ushort*) info, wcslen(info));
+                                // qDebug() << "deleteShortcuts " << targetPath << " " <<
+                                //        instPath;
+                                if (WPMUtils::isUnder(targetPath,
+                                                      instPath)) {
+                                    QFile::remove(path);
+                                    // qDebug() << "deleteShortcuts true" << ok;
+                                }
                             }
                         }
+                        ppf->Release();
                     }
-                    ppf->Release();
                 }
             }
         }
